change 'entity.User' to 'model.User' to account for refactoring changes in main
