+++ src/main/java/model/Image.java
@@ -1,4 +1,4 @@
-package entity;
+package model;
 
 import java.io.File;
